requirements.txt
openai-whisper==20250625
python-dotenv==1.1.1
torch>=2.0.0
numpy>=1.21.0
tqdm>=4.60.0